Split the MainActivity.kt
